requirements.txt
biopython==1.85
pandas>=1.26.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.13.0
numpy<2        # <--- important! downgrade from 2.0
psutil
scikit-learn
biotite

fastapi==0.111.0
uvicorn[standard]==0.29.0
celery[redis]==5.4.0
pydantic==2.7.2
python-multipart==0.0.9
